+++ CHANGES.rst
@@ -2,6 +2,13 @@
 Changes
 =======
 
+development version
+-------------------
+
+* :issue:`393`: Using ``--info-file`` no longer crashes when processing
+  paired-end data. However, the info file itself will only contain results
+  for R1.
+
 v2.4 (2019-07-09)
 -----------------